Disc Inventory
added a 2 new classes: Cd and Dvd.
Now my inventory is a Disc inventory. Both new classes extend the Product class

Changed the addToInventory and method to add a Cd or a Dvd. now you have to choose one option before the user's imputs to the array

Now you can't add quantity to a product when its status is "Desativado"