Renovable arrays
Now you can add more Products without having to remove the others

moved the first array and added a second to the method execMenu, this way I don't need adicional parameters.

When adding a second inventory, the array bigger goes to the new one that has the size of (theBiggerArray.lenght + numberOfNewItems)

changed every method so they can return accordingly to the smallest an bigger array.

added a static int to the class product that represents the total number of Instances, this way enumerating becomes easier